Unify next_channel_from_routes for initiator and mediator
Both the initator and mediator had a function next_channel_from_routes.
This commit unifies them in the channel module.
